mv warmup_lr_lambda
(cherry picked from commit e487737c080eb3084ec94533ae2fb4d5c7ed98c7)
