+++ src/fairchem/core/modules/scheduler.py
@@ -9,10 +9,33 @@ from __future__ import annotations
 
 import inspect
 import math
+from bisect import bisect
 
 import torch.optim.lr_scheduler as lr_scheduler
 
-from fairchem.core.common.utils import warmup_lr_lambda
+
+def warmup_lr_lambda(current_step: int, optim_config):
+    """Returns a learning rate multiplier.
+    Till `warmup_steps`, learning rate linearly increases to `initial_lr`,
+    and then gets multiplied by `lr_gamma` every time a milestone is crossed.
+    """
+
+    # keep this block for older configs that have warmup_epochs instead of warmup_steps
+    # and lr_milestones are defined in epochs
+    if (
+        any(x < 100 for x in optim_config["lr_milestones"])
+        or "warmup_epochs" in optim_config
+    ):
+        raise Exception(
+            "ConfigError: please define lr_milestones in steps not epochs and define warmup_steps instead of warmup_epochs"
+        )
+
+    if current_step <= optim_config["warmup_steps"]:
+        alpha = current_step / float(optim_config["warmup_steps"])
+        return optim_config["warmup_factor"] * (1.0 - alpha) + alpha
+    else:
+        idx = bisect(optim_config["lr_milestones"], current_step)
+        return pow(optim_config["lr_gamma"], idx)
 
 
 class CosineLRLambda:
